fix opening conversation from archive
