+++ wire-ios/Wire-iOS/Sources/UserInterface/ConversationList/ArchivedList/ArchivedListViewControllerDelegate.swift
@@ -19,6 +19,5 @@
 import WireDataModel
 
 protocol ArchivedListViewControllerDelegate: AnyObject {
-    func archivedListViewControllerWantsToDismiss(_ controller: ArchivedListViewController)
-    func archivedListViewController(_ controller: ArchivedListViewController, didSelectConversation conversation: ZMConversation)
+    func archivedListViewController(_ viewController: ArchivedListViewController, didSelectConversation conversation: ZMConversation)
 }
