+++ wire-ios/Wire-iOS/Sources/UserInterface/ConversationList/Container/ConversationListViewController+State.swift
@@ -20,6 +20,7 @@ import UIKit
 
 extension ConversationListViewController {
 
+    #warning("TODO: remove")
     func setState(
         _ state: ConversationListState,
         animated: Bool,
